fix for hobo:subsite generator

+++ hobo/lib/generators/hobo/subsite.rb
@@ -20,7 +20,7 @@ module Generators
           end
           say "Renaming app/views/taglibs/application.dryml to app/views/taglibs/front_site.dryml" \
               unless options[:quiet]
-          FileUtils.mv('app/views/taglibs/application.dryml', "app/views/taglibs/front_site.dryml")
+          FileUtils.mv('app/views/taglibs/application.dryml', "app/views/taglibs/front_site.dryml") unless options[:pretend]
           copy_file "application.dryml", 'app/views/taglibs/application.dryml'
         end
